Move SessionId and KeySetId to more appropriate packages.
