+++ core/src/main/java/com/novoda/noplayer/drm/DownloadedModularDrm.java
@@ -1,5 +1,7 @@
 package com.novoda.noplayer.drm;
 
+import com.novoda.noplayer.model.KeySetId;
+
 public interface DownloadedModularDrm extends DrmHandler {
 
     KeySetId getKeySetId();
