+++ core/src/main/java/com/novoda/noplayer/exoplayer/drm/FrameworkDrmSession.java
@@ -2,7 +2,6 @@ package com.novoda.noplayer.exoplayer.drm;
 
 import com.google.android.exoplayer2.drm.DrmSession;
 import com.google.android.exoplayer2.drm.FrameworkMediaCrypto;
-import com.novoda.noplayer.drm.SessionId;
 
 interface FrameworkDrmSession extends DrmSession<FrameworkMediaCrypto> {
