+++ core/src/main/java/com/novoda/noplayer/exoplayer/drm/LocalDrmSession.java
@@ -3,8 +3,7 @@ package com.novoda.noplayer.exoplayer.drm;
 import android.support.annotation.Nullable;
 
 import com.google.android.exoplayer2.drm.FrameworkMediaCrypto;
-import com.novoda.noplayer.drm.KeySetId;
-import com.novoda.noplayer.drm.SessionId;
+import com.novoda.noplayer.model.KeySetId;
 
 import java.util.Collections;
 import java.util.HashMap;
