+++ core/src/main/java/com/novoda/noplayer/model/KeySetId.java
@@ -1,4 +1,4 @@
-package com.novoda.noplayer.drm;
+package com.novoda.noplayer.model;
 
 import java.util.Arrays;
